refactor: extract useUserPreferences from App.tsx (phase 4b)
Moves theme/systemDark/preferredCurrency/fxRate state and three effects
(media-query listener, dark-class applier, Supabase prefs persist) into
src/hooks/useUserPreferences.ts.

Hook takes session + prefsLoadedRef (defined in App.tsx so the single ref
is still shared by syncTransactions). Returns flat setters so syncTransactions,
handleSignOut, and handleResetAllData can call them without wrappers.

App.tsx: 750 → 717 lines. Behavior and all 658 tests unchanged.
